+++ stdlib/rand.h
@@ -1,0 +1,6 @@
+#pragma once
+
+void srand(unsigned int seed);
+int rand(void);
+int rand_range(int min, int max);
+int rand_bool(void);