+++ stdlib/time.c
@@ -1,0 +1,5 @@
+#include <time.h>
+
+long now(void) {
+    return time(NULL);
+}